pyproject.toml
[project]
name = "xch-mind"
version = "0.1.0"
description = "Multi-agent Interpretive Nexus Discovery for CH Linked Data"
authors = [
    {name = "Stingo Michele", email = "stingomichele@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-ollama>=0.2.0",
    "langchain-core>=0.3.0",
    "rdflib>=7.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "lxml>=5.0.0",
    "spacy>=3.7.0",
    "pyshacl>=0.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
